+++ doc/CHANGELOG.md
@@ -1,1 +1,13 @@
 ﻿## Change log
+
+### V1.0.0 -> V1.1.0
+
+- Main new features
+  - OSD show local time
+  - FullScreen progress indicator
+  - Quick start mode
+- Other changes
+  - Update libmpv and Qt
+  - Update translations
+  - Fix some bugs
+  - Minor tweaks